+++ src/js/components/Tabs/stories/Plain.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import UncontrolledTabs from './Uncontrolled';
+
+storiesOf('Tabs', module).add('Plain', () => <UncontrolledTabs plain />);
